+++ Giunne/android/src/main/AndroidManifest.xml
@@ -12,6 +12,7 @@
         android:allowBackup="false"
         android:supportsRtl="true"
         android:largeHeap="true"
+        android:usesCleartextTraffic="true"
         android:theme="@style/Theme.AppCompat.Light.NoActionBar">
 
         <activity
