+++ Giunne/common/src/commonMain/kotlin/com/project/giunne/common/presentation/community/teacher/TeacherCommunityScreen.kt
@@ -26,10 +26,9 @@ import com.project.giunne.common.presentation.certification.student.state.CertPa
 import com.project.giunne.common.presentation.common.addFocusCleaner
 import com.project.giunne.common.presentation.common.content.Loader
 import com.project.giunne.common.presentation.common.scrollbar.VerticalScrollbar
-import com.project.giunne.common.presentation.common.text.GPText
-import com.project.giunne.common.presentation.community.student.content.CommunityItemRow
-import com.project.giunne.common.presentation.community.student.content.SearchRow
-import com.project.giunne.common.presentation.community.student.content.SelectableDialog
+import com.project.giunne.common.presentation.community.content.CommunityItemRow
+import com.project.giunne.common.presentation.community.content.SearchRow
+import com.project.giunne.common.presentation.community.content.SelectableDialog
 import com.project.giunne.common.presentation.community.student.dummy.CommunityDto
 import com.project.giunne.common.presentation.community.student.dummy.roadmapCommunityList
 import com.project.giunne.common.presentation.community.student.dummy.runningCommunityList
@@ -37,7 +36,6 @@ import com.project.giunne.common.presentation.community.student.state.DatePriori
 import com.project.giunne.common.ui.theme.GPColor
 import com.project.giunne.common.util.GLog
 import com.project.giunne.common.util.gdp
-import com.project.giunne.common.util.gsp
 import org.jetbrains.compose.resources.painterResource
 
 private const val TAG = "TeacherCommunityScreen"
